refactor vectorDB and embed first
