+++ embed/__init__.py
@@ -1,0 +1,2 @@
+from .embedding import Embedding, EmbeddingType
+from .embedded_files_cache import EmbeddedFilesCache
