+++ options/__init__.py
@@ -1,1 +1,1 @@
-from config import Options, ChromaOptions, PineconeOptions
+from .config import Options, ChromaOptions, PineconeOptions
